Revert from precompiling cryptoauthlib in order to reduce flash usage

+++ scripts/make.sh
@@ -1,18 +1,14 @@
 #!/bin/bash
 
 SCRIPTPATH="$( cd "$(dirname "$0")" ; pwd -P )"
+TARGET=$SCRIPTPATH/../examples/http/http.ino
+BOARD_CONFIG="DxCore:megaavr:avrdb:appspm=no,chip=avr128db48,clock=24internal,bodvoltage=1v9,bodmode=disabled,eesave=enable,resetpin=reset,millis=tcb2,startuptime=8,wiremode=mors2"
 
-TARGET=$SCRIPTPATH/../examples/sandbox/sandbox.ino
-EXTRA_ARGS=--clean
-
-if [ "$1" = "clean" ]; then
-    EXTRA_ARGS=--clean
+if [ -d "$SCRIPTPATH/../build" ]; then
+    rm -r "$SCRIPTPATH/../build"
 fi
 
-rm -r "$SCRIPTPATH/../build"
+echo "Compiling..."
 
-arduino-cli compile \
-                "$TARGET" \
-                $EXTRA_ARGS \
-                -b DxCore:megaavr:avrdb:appspm=no,chip=avr128db48,clock=24internal,bodvoltage=1v9,bodmode=disabled,eesave=enable,resetpin=reset,millis=tcb2,startuptime=8,wiremode=mors2 \
-                --output-dir "$SCRIPTPATH/../build"
+# Extra args appended to the end, e.g. --clean by the $1 flag
+arduino-cli compile "$TARGET" -b $BOARD_CONFIG --output-dir "$SCRIPTPATH/../build" $1
